feat(BreakdownSearchScene): move filtering to ByFrameRepeater
