+++ src/Components/ServiceScene/Breakdowns/BreakdownSearchScene.tsx
@@ -1,12 +1,9 @@
 import { SceneComponentProps, SceneObjectBase, SceneObjectState } from '@grafana/scenes';
 import React, { ChangeEvent } from 'react';
 import { ByFrameRepeater } from './ByFrameRepeater';
-import { DataFrame } from '@grafana/data';
 import { SearchInput } from './SearchInput';
 import { LabelBreakdownScene } from './LabelBreakdownScene';
 import { FieldsBreakdownScene } from './FieldsBreakdownScene';
-import { fuzzySearch } from '../../../services/search';
-import { getLabelValueFromDataFrame } from 'services/levels';
 
 export interface BreakdownSearchSceneState extends SceneObjectState {
   filter?: string;
@@ -17,6 +14,7 @@ export class BreakdownSearchScene extends SceneObjectBase<BreakdownSearchSceneSt
     super({
       filter: '',
     });
+    console.log('new search');
   }
 
   public static Component = ({ model }: SceneComponentProps<BreakdownSearchScene>) => {
@@ -46,30 +44,9 @@ export class BreakdownSearchScene extends SceneObjectBase<BreakdownSearchSceneSt
       const body = this.parent.state.body;
       body?.forEachChild((child) => {
         if (child instanceof ByFrameRepeater && child.state.body.isActive) {
-          let haystack: string[] = [];
-
-          child.iterateFrames((frames, seriesIndex) => {
-            const labelValue = getLabelValue(frames[seriesIndex]);
-            haystack.push(labelValue);
-          });
-          fuzzySearch(haystack, filter, (data) => {
-            if (data && data[0]) {
-              // We got search results
-              child.filterFrames((frame: DataFrame) => {
-                const label = getLabelValue(frame);
-                return data[0].includes(label);
-              });
-            } else {
-              // reset search
-              child.filterFrames(() => true);
-            }
-          });
+          child.filterByString(filter);
         }
       });
     }
   }
-}
-
-export function getLabelValue(frame: DataFrame) {
-  return getLabelValueFromDataFrame(frame) ?? 'No labels';
 }
